+++ src/onegov.agency/onegov/agency/pdf.py
@@ -1,13 +1,25 @@
 from datetime import date
 from io import BytesIO
+from onegov.core.utils import module_path
 from onegov.pdf import page_fn_footer
 from onegov.pdf import page_fn_header_and_footer
+from onegov.pdf import page_fn_header_logo
+from onegov.pdf import page_fn_header_logo_and_footer
 from onegov.pdf import Pdf
+from os import path
 from reportlab.lib.units import cm
 
 
-class DefaultAgencyPdf(Pdf):
+class AgencyPdfDefault(Pdf):
     """ A standard PDF of an agency. """
+
+    @property
+    def page_fn(self):
+        return page_fn_footer
+
+    @property
+    def page_fn_later(self):
+        return page_fn_header_and_footer
 
     @classmethod
     def from_agencies(cls, agencies, author, title, toc, exclude):
@@ -21,9 +33,11 @@ class DefaultAgencyPdf(Pdf):
             author=author
         )
         pdf.init_a4_portrait(
-            page_fn=page_fn_footer,
-            page_fn_later=page_fn_header_and_footer
+            page_fn=pdf.page_fn,
+            page_fn_later=pdf.page_fn_later
         )
+        pdf.spacer()
+        pdf.spacer()
         pdf.h(title)
         pdf.spacer()
         if toc:
@@ -32,7 +46,7 @@ class DefaultAgencyPdf(Pdf):
         for agency in agencies:
             if agency.is_hidden_from_public:
                 continue
-            pdf.agency(agency, exclude)
+            pdf.agency(agency, exclude, skip_title=title == agency.title)
             pdf.pagebreak()
         pdf.generate()
 
@@ -78,7 +92,8 @@ class DefaultAgencyPdf(Pdf):
         if data:
             self.table(data, [5.5 * cm, 0.5 * cm, None])
 
-    def agency(self, agency, exclude, level=1, content_so_far=False):
+    def agency(self, agency, exclude, level=1, content_so_far=False,
+               skip_title=False):
         """ Adds a single agency with the portrait and memberships. """
 
         if (level < 4) and content_so_far:
@@ -86,8 +101,9 @@ class DefaultAgencyPdf(Pdf):
         else:
             self.spacer()
 
-        self.h(agency.title, level)
-        self.story[-1].keepWithNext = True
+        if not skip_title:
+            self.h(agency.title, level)
+            self.story[-1].keepWithNext = True
 
         has_content = False
         if agency.portrait:
@@ -113,3 +129,68 @@ class DefaultAgencyPdf(Pdf):
             has_content = has_content or child_has_content
 
         return has_content
+
+
+class AgencyPdfZg(AgencyPdfDefault):
+    """ A PDF with the CI of the canton of ZG. """
+
+    @staticmethod
+    def page_fn_footer(canvas, doc):
+        """ A footer with the title and print date on the left and the page
+        numbers on the right. """
+
+        canvas.saveState()
+        canvas.setFont('Helvetica', 9)
+
+        canvas.drawString(
+            doc.leftMargin,
+            doc.bottomMargin / 2 + 12,
+            doc.title
+        )
+        canvas.drawString(
+            doc.leftMargin,
+            doc.bottomMargin / 2,
+            f'Druckdatum: {doc.created}'
+        )
+        canvas.drawRightString(
+            doc.pagesize[0] - doc.rightMargin,
+            doc.bottomMargin / 2,
+            f'{canvas._pageNumber}'
+        )
+        canvas.restoreState()
+
+    @property
+    def page_fn(self):
+        return page_fn_header_logo
+
+    @property
+    def page_fn_later(self):
+        return self.page_fn_footer
+
+    def __init__(self, *args, **kwargs):
+        filename = path.join(
+            module_path('onegov.agency', 'static/logos'),
+            'canton-zg-bw.svg'
+        )
+        with open(filename) as file:
+            logo = file.read()
+        kwargs['logo'] = logo
+        super(AgencyPdfDefault, self).__init__(*args, **kwargs)
+
+
+class AgencyPdfAr(AgencyPdfDefault):
+    """ A PDF with the CI of the canton of AR. """
+
+    @property
+    def page_fn(self):
+        return page_fn_header_logo_and_footer
+
+    def __init__(self, *args, **kwargs):
+        filename = path.join(
+            module_path('onegov.agency', 'static/logos'),
+            'canton-ar.svg'
+        )
+        with open(filename) as file:
+            logo = file.read()
+        kwargs['logo'] = logo
+        super(AgencyPdfDefault, self).__init__(*args, **kwargs)